+++ packages/core/src/index.ts
@@ -1,2 +1,3 @@
+export * from "./crossCheck";
 export * from "./schemas";
 export * from "./syntheticExamples";